requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
groq>=0.9.0
upstash-vector>=0.5.0
sentence-transformers>=3.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pydantic>=2.0.0